Export RichTextEditFindBar headers
The PlainTextEditFindBar is already exported, just do the same for
RichTextEditFindBar.

I am currently porting KJots to KPIMTextEdit, as it seems to be much
nicer compared to KTextWidgets. One of the features I like is inline search,
which I would like to have both for KJotsEdit (which can be RichTextComposer)
and KJotsBrowser (which is QTextBrowser). Since QTextBrowser is a
subclass of QTextEdit, having RichTextEditFindBar there would be useful :)